⚡ Bolt: Optimize Dashboard Calendar Grid Rendering
💡 What: Refactored the inline calendar grid rendering logic into a memoized `DayColumn` component.
🎯 Why: The original implementation re-calculated event layouts (partitioning, grouping, positioning) for every day on every render of the `Dashboard` component. This caused significant performance overhead, especially when unrelated state (like toggling tasks or hovering events) changed.
📊 Impact: Reduces unnecessary re-renders of heavy day columns. Layout calculations now only run when events or days change.
🔬 Measurement: Verified with unit tests and visual inspection via Playwright screenshot.
